HPath.IO.Utils

Description

Random and general IO/monad utilities.

Documentation

whenM :: Monad m => m Bool -> m () -> m ()

If the value of the first argument is True, then execute the action provided in the second argument, otherwise do nothing.

unlessM :: Monad m => m Bool -> m () -> m ()

If the value of the first argument is False, then execute the action provided in the second argument, otherwise do nothing.