System.Posix.Directory.Foreign

Documentation

newtype DirType

Constructors

DirType Int

data Flags

Constructors

Flags Int

UnsupportedFlag String

unFlags :: Flags -> Int

isSupported :: Flags -> Bool

Returns True if posix-paths was compiled with support for the provided flag. (As of this writing, the only flag for which this check may be necessary is oCloexec; all other flags will always yield True.)

oCloexec :: Flags

O_CLOEXEC is not supported on every POSIX platform. Use isSupported oCloexec to determine if support for O_CLOEXEC was compiled into your version of posix-paths. (If not, using oCloexec will throw an exception.)

dtBlk :: DirType

oAppend :: Flags

dtChr :: DirType

pathMax :: Int

oAsync :: Flags

dtDir :: DirType

oCreat :: Flags

dtFifo :: DirType

unionFlags :: [Flags] -> CInt

dtLnk :: DirType

oDirectory :: Flags

oExcl :: Flags

dtReg :: DirType

oNoctty :: Flags

dtSock :: DirType

oNofollow :: Flags

dtUnknown :: DirType

oNonblock :: Flags

oRdonly :: Flags

oWronly :: Flags

oRdwr :: Flags

oSync :: Flags

oTrunc :: Flags